requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
Pygments==2.15.1
python-nmap==0.7.1
requests==2.28.2
rich==13.3.4
urllib3==1.26.15
